Split out downloader from main.c into its own file.
